+++ studio/src/app/utils/editor/firestore.utils.tsx
@@ -1,3 +1,6 @@
+import firebase from 'firebase/app';
+import 'firebase/firestore';
+
 export class FirestoreUtils {
   static filterDelete<T>(obj: T, replaceWithNull: boolean = false): T {
     if (typeof obj !== 'object' || Array.isArray(obj)) {
@@ -30,10 +33,6 @@ export class FirestoreUtils {
       return false;
     }
 
-    const firestoreDelete = Object.keys(attr).find((key: string) => {
-      return attr[key] === 'FieldValue.delete';
-    });
-
-    return firestoreDelete !== null && firestoreDelete !== undefined;
+    return JSON.stringify(attr) === JSON.stringify(firebase.firestore.FieldValue.delete());
   }
 }